backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-multipart==0.0.18
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]==2.5.3
python-dotenv==1.0.1
tzdata==2025.2
alembic==1.13.1
pytest==8.0.0
bcrypt==4.0.1
boto3==1.34.5
httpx==0.27.0